+++ deployment/v1_7_0/changesets/configure_chains_for_lanes_from_topology.go
@@ -29,7 +29,7 @@ type CommitteeVerifierRemoteChainConfig struct {
 	RemovedAllowlistedSenders []string
 	FeeUSDCents               uint16
 	GasForVerification        uint32
-	PayloadSizeBytes          uint32
+	PayloadSizeBytes          uint16
 }
 
 type CommitteeVerifierInputConfig struct {
